Cleanup
The files now are saved/deleted only if a file was send or if a product has a file.

+++ server/src/controlers/products.ts
@@ -2,13 +2,13 @@
 import mongoose from 'mongoose';
 import ProductMessage from '../models/product-schema.js';
 import { FileManager } from '../lib/FileManager.js';
-import path from 'path';
-import { dirname } from 'path';
-import { fileURLToPath } from 'url';
+// import path from 'path';
+// import { dirname } from 'path';
+// import { fileURLToPath } from 'url';
 
 
 
-const __dirname = dirname( fileURLToPath( import.meta.url ) );
+// const __dirname = dirname( fileURLToPath( import.meta.url ) );
 
 interface Product
 {
@@ -40,21 +40,30 @@ export const addProduct = async ( req: any, res: any ) =>
 {
     const product = generateSrclessProduct( req.body );
     const files = req.files;
-    let file = files[ 0 ];
-    const fileName = Date.now();
-    const filePath = `uploads/${ fileName }.png`;
+    const file = files[ 0 ];
+    let fileName;
+    let filePath;
+    if ( file )
+    {
+        fileName = Date.now();
+        filePath = `uploads/${ fileName }.png`;
+        const src = `http://localhost:5000/${ filePath }`;
+        product.src = src.replace( /\\/g, "/" );
+    }
 
-    product.src = `http://localhost:5000/${ filePath }`;
-
-
-    product.src = product.src.replace( /\\/g, "/" );
 
     try
     {
         const newProduct = new ProductMessage( product );
         console.log( 'New Product: ' + newProduct );
+        //saving the product to the db
         await newProduct.save();
-        saveFile( `public/${ filePath }`, file.buffer );
+        //saving the file to the uploads folder
+        if ( file )
+        {
+            saveFile( `public/${ filePath }`, file.buffer );
+        }
+
         res.status( 201 ).json( { status: 'success', product: newProduct, message: "Product added succesfully" } );
     } catch ( error )
     {
@@ -67,7 +76,6 @@ export const addProduct = async ( req: any, res: any ) =>
 export const addProducts = async ( req: any, res: any ) =>
 {
     const proudcts = req.body;
-    const files = req.files;
 
 
     for ( let product of proudcts )
@@ -82,8 +90,6 @@ export const addProducts = async ( req: any, res: any ) =>
             res.status( 409 ).json( { status: 'fail', message: 'Unsuccesfull save! Product name duplicate' } );
         }
     }
-
-
 };
 
 
@@ -132,7 +138,8 @@ export const deleteProduct = async ( req: any, res: any ) =>
     const selectedProduct = await ProductMessage.findById( _id );
     await ProductMessage.findByIdAndDelete( _id );
 
-    FileManager.delete( getPathToDelete( selectedProduct.src ) );
+    if ( selectedProduct.src )
+        FileManager.delete( getPathToDelete( selectedProduct.src ) );
 
     res.status( 201 ).send( { status: 'success', message: `Item ${ _id } was deleted` } );
